requirements.txt
PyMuPDF==1.23.4
PySide2==5.15.2.1
pyyaml==6.0